test (LottoTest): 비즈니스 로직 이동 LottoDto -> Lotto
