+++ src/main/java/lotto/domain/Lotto.java
@@ -1,6 +1,8 @@
 package lotto.domain;
 
 import java.util.List;
+import lotto.domain.dto.BonusNumberDto;
+import lotto.domain.dto.WinningLottoDto;
 import lotto.validator.LottoValidator;
 
 public class Lotto {
@@ -17,6 +19,21 @@ public class Lotto {
 
     private List<Integer> sort(List<Integer> numbers) {
         return numbers.stream().sorted().toList();
+    }
+
+    public int matchWinningLottoCount(final WinningLottoDto winningLottoDto) {
+        int matchCount = 0;
+
+        for (Integer number : numbers) {
+            matchCount += (int) winningLottoDto.numbers().stream().filter(winningNumber -> winningNumber.equals(number))
+                    .count();
+        }
+
+        return matchCount;
+    }
+
+    public boolean hasBonusNumber(final BonusNumberDto bonusNumberDto) {
+        return numbers.stream().anyMatch(i -> i.equals(bonusNumberDto.number()));
     }
 
     public List<Integer> getNumbers() {
