+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,9 +1,14 @@
 package lotto.domain;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
+import lotto.domain.dto.BonusNumberDto;
+import lotto.domain.dto.WinningLottoDto;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 class LottoTest {
@@ -20,4 +25,78 @@ class LottoTest {
     }
 
     // 아래에 추가 테스트 작성 가능
+
+    @Nested
+    @DisplayName("matchWinningLottoCount 메소드 test")
+    class MatchWinningLottoCount {
+        private Lotto lotto;
+
+        @BeforeEach
+        void beforeEach() {
+            lotto = new Lotto(List.of(4, 8, 17, 26, 39, 40));
+        }
+
+        @DisplayName("당첨 번호와 발행된 로또를 비교하여 일치하는 숫자가 있다면 일치하는 개수를 반환한다.")
+        @Test
+        void matching_numbers_present() {
+            // given
+            WinningLottoDto winningLottoDto = new WinningLottoDto(List.of(4, 8, 17, 30, 41, 45));
+
+            // when
+            int matchedWinningLottoCount = lotto.matchWinningLottoCount(winningLottoDto);
+
+            // then
+            assertThat(matchedWinningLottoCount).isEqualTo(3);
+        }
+
+        @DisplayName("당첨 번호와 발행된 로또를 비교하여 일치하는 숫자가 없다면 0을 반환한다")
+        @Test
+        void matching_numbers_not_present() {
+            // given
+            WinningLottoDto winningLottoDto = new WinningLottoDto(List.of(1, 19, 23, 33, 42, 45));
+
+            // when
+            int matchedWinningLottoCount = lotto.matchWinningLottoCount(winningLottoDto);
+
+            // then
+            assertThat(matchedWinningLottoCount).isEqualTo(0);
+        }
+    }
+
+    @Nested
+    @DisplayName("hasBonusNumber 메소드 test")
+    class HasBonusNumber {
+        private Lotto lotto;
+
+        @BeforeEach
+        void beforeEach() {
+            lotto = new Lotto(List.of(4, 8, 17, 26, 39, 40));
+        }
+
+        @DisplayName("발행된 로또에 보너스 숫자가 있다면 true 반환")
+        @Test
+        void Lotto_contain_bonus_number() {
+            // given
+            BonusNumberDto bonusNumberDto = new BonusNumberDto(45);
+
+            // when
+            boolean hasBonusNumber = lotto.hasBonusNumber(bonusNumberDto);
+
+            // then
+            assertThat(hasBonusNumber).isEqualTo(false);
+        }
+
+        @DisplayName("발행된 로또에 보너스 숫자가 없다면 false 반환")
+        @Test
+        void Lotto_not_contain_bonus_number() {
+            // given
+            BonusNumberDto bonusNumberDto = new BonusNumberDto(4);
+
+            // when
+            boolean hasBonusNumber = lotto.hasBonusNumber(bonusNumberDto);
+
+            // then
+            assertThat(hasBonusNumber).isEqualTo(true);
+        }
+    }
 }
